refactored to single Block Cart
