--- blocks/commerce-cart-order-summary/commerce-cart-order-summary.css
@@ -1,1 +1,0 @@
-/* stylelint-disable no-empty-source */
