--- blocks/commerce-cart-summary/commerce-cart-summary.css
@@ -1,1 +1,0 @@
-/* stylelint-disable no-empty-source */
