+++ blocks/commerce-cart/commerce-cart.css
@@ -1,1 +1,26 @@
-/* stylelint-disable no-empty-source */
+.cart__wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: var(--grid-4-gutters);
+  padding: var(--spacing-large) 0;
+}
+
+.cart__wrapper > .cart-empty-cart {
+  flex: 1;
+} 
+
+@media (width >= 900px) {
+  .cart__wrapper {
+    flex-direction: row;
+  }
+
+  .cart__left-column {
+    flex: 1;
+    flex-basis: 70%;
+  }
+
+  .cart__right-column {
+    flex: 1;
+    flex-basis: 30%;
+  }
+}
