--- blocks/commerce-empty-cart/commerce-empty-cart.css
@@ -1,1 +1,0 @@
-/* stylelint-disable no-empty-source */
